# crates/api-gateway/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Denial Navigator API",
    "version": "1.0.0",
    "description": "Self-hosted denial-management API (Rust / axum rewrite). Every browser request goes through this gateway; it enforces authentication, role-based authorisation, per-request account currency and HIPAA access auditing.\n\nAuth: `POST /api/v1/auth/login` returns a bearer token. When TOTP is required it returns an mfa-scope token usable only on the `/auth/totp/*`, `/auth/login/totp` and `/auth/me` paths until `POST /auth/login/totp` upgrades it. Sibling services present `X-Service-Key` instead."
  },
  "servers": [
    {
      "url": "/",
      "description": "This gateway"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "serviceKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Service-Key"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid credentials.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but not allowed.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "No such resource.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/v1/analyses": {
      "get": {
        "summary": "List AI analyses",
        "tags": [
          "analyses"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "denial_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "claim_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Max rows to return."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/analyses/generate": {
      "post": {
        "summary": "Run the full pipeline for a denial (RAG + LLM)",
        "tags": [
          "analyses"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Rate limited per caller (ANALYSES_RATE_LIMIT/min).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "denial_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "temperature": {
                    "type": "number",
                    "default": 0.3
                  }
                },
                "required": [
                  "denial_id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "description": "Rate limited."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/analyses/generate-jobs": {
      "post": {
        "summary": "Queue asynchronous recommendation generation",
        "tags": [
          "analyses"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Returns a durable job ID; poll its status endpoint for the result.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "denial_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "temperature": {
                    "type": "number",
                    "default": 0.3
                  }
                },
                "required": [
                  "denial_id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Queued",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "429": {
            "description": "Rate limited."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/analyses/generate-jobs/{job_id}": {
      "get": {
        "summary": "Get asynchronous recommendation job status",
        "tags": [
          "analyses"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/analyses/store": {
      "post": {
        "summary": "Store an analysis result (usually called by the LLM service)",
        "tags": [
          "analyses"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "serviceKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "denial_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "claim_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "model_name": {
                    "type": "string"
                  },
                  "raw_prompt": {
                    "type": "string"
                  },
                  "raw_response": {
                    "type": "string"
                  },
                  "parsed_result": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "prompt_tokens": {
                    "type": "integer"
                  },
                  "completion_tokens": {
                    "type": "integer"
                  },
                  "total_tokens": {
                    "type": "integer"
                  }
                },
                "required": [
                  "denial_id",
                  "claim_id",
                  "model_name",
                  "raw_prompt",
                  "raw_response",
                  "parsed_result"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  }
}